fix: organize exception files and fixes the imports
