+++ app/exceptions/client_exc.py
